requirements.txt
selenium
undetected-chromedriver
pandas
python-dateutil
holidays